+++ src/main/kotlin/org/jetbrains/changelog/extensions.kt
@@ -1,0 +1,8 @@
+package org.jetbrains.changelog
+
+import groovy.lang.Closure
+
+fun <T : Any> closure(function: () -> T) = object : Closure<T>(null) {
+    @Suppress("unused")
+    fun doCall() = function()
+}